fix compilation post API changes

+++ crates/core/keys/src/keys/ivk.rs
@@ -129,7 +129,7 @@ impl IncomingViewingKeyVar {
         //
         // Constrain: ivk_mod_q = mod_r * a + ivk_mod_r
         let mod_r_var = FqVar::new_constant(cs.clone(), r_modulus)?;
-        let a_var = FqVar::new_witness(cs.clone(), || Ok(Fq::from(a)))?;
+        let a_var = FqVar::new_witness(cs.clone(), || Ok(Fq::from(a as u64)))?;
         let rhs = &mod_r_var * &a_var + &ivk_mod_r;
         ivk_mod_q.enforce_equal(&rhs)?;
